refactor(setup): extract SetupViewModel from SetupView
Move all setup state and logic (server validation, auth detection, classic/
token/device-flow login handlers, error parsing) into an @Observable
SetupViewModel, matching the app's MVVM pattern. SetupView keeps only view
composition. No behavior change. SetupView 1440->957 lines.
