+++ romm/romm/UI/App/SetupView.swift
@@ -15,85 +15,16 @@ enum SetupField: Hashable {
 
 struct SetupView: View {
     let appViewModel: AppViewModel
-    @State private var serverURL = ""
-    @State private var username = ""
-    @State private var password = ""
-    @State private var showPassword = false
-    @State private var showConnectionDetails = false
-
-    // Connection validation states
-    @State private var isConnecting = false
-    @State private var serverValidated = false
-    @State private var detectedServerVersion: String?
-    @State private var connectionError: String?
-    @State private var connectionErrorDetails: String?
-    @State private var isErrorExpanded = false
-    @State private var isVersionWarning = false
-    @State private var didAcceptIncompatibleVersion = false
-
-    // Version warning detail text
-    @State private var versionWarningTitle: String?
-    @State private var versionWarningDetails: String?
-
-    // Auth capability detection
-    @State private var detectedAuthCapability: HeartbeatRepository.AuthCapabilities?
-    @State private var selectedAuthMethod: AuthMethod = .classic
-
-    // Client Token states
-    @State private var clientTokenInput = ""
-    @State private var isExchangingToken = false
-    @State private var clientTokenError: String?
-    @State private var showQRScanner = false
-
-    // Device flow (browser) states
-    @State private var deviceFlowInit: DeviceAuthInitResponse?
-    @State private var isDeviceFlowRunning = false
-    @State private var deviceFlowError: String?
-    @State private var deviceFlowTask: Task<Void, Never>?
-
-    // "Other sign-in options" disclosure
-    @State private var showOtherOptions = false
+    @State private var viewModel: SetupViewModel
 
     // Focus
     @FocusState private var focusedField: SetupField?
 
     private var connectionLogger: ConnectionLogger { ConnectionLogger.shared }
-    private let launchArguments = ProcessInfo.processInfo.arguments
-    private var shouldShowLoginForUITests: Bool { launchArguments.contains("-ui_testing_show_login") }
 
-    // MARK: - Computed helpers
-
-    private var isBusy: Bool {
-        appViewModel.appData.isLoading || isExchangingToken || isConnecting || isDeviceFlowRunning
-    }
-
-    /// Whether the browser device-flow is the primary path. Optimistic before the
-    /// server is validated; after validation, only when the server supports it.
-    private var showDeviceFlowPrimary: Bool {
-        guard let cap = detectedAuthCapability else { return true }
-        return cap.deviceFlow
-    }
-
-    /// Fallback methods (username/password, API token) shown under "Other options".
-    private var otherMethods: [AuthMethod] {
-        guard let cap = detectedAuthCapability else { return [.classic, .clientToken] }
-        return AuthMethod.availableMethods(for: cap).filter { $0 != .deviceFlow }
-    }
-
-    private var canSubmit: Bool {
-        guard !isBusy && !serverURL.isEmpty else { return false }
-        switch selectedAuthMethod {
-        case .deviceFlow:
-            return false // Browser sign-in has its own button, not this one.
-        case .classic:
-            return !username.isEmpty && !password.isEmpty
-        case .clientToken:
-            return !clientTokenInput.isEmpty
-        }
-    }
-
-    private var hasLoginError: Bool {
-        appViewModel.appData.errorMessage != nil
+    init(appViewModel: AppViewModel) {
+        self.appViewModel = appViewModel
+        self._viewModel = State(wrappedValue: SetupViewModel(appViewModel: appViewModel))
     }
 
     // MARK: - Body
@@ -153,7 +84,7 @@ struct SetupView: View {
                     if !connectionLogger.logs.isEmpty || connectionLogger.isConnecting {
                         ConnectionDebugPanel(
                             logs: connectionLogger.logs,
-                            isExpanded: $showConnectionDetails
+                            isExpanded: $viewModel.showConnectionDetails
                         )
                         .padding(.horizontal, 20)
                         .padding(.top, 16)
@@ -169,33 +100,33 @@ struct SetupView: View {
         .preferredColorScheme(.dark)
         .tint(SetupTheme.coralLight)
         .onAppear {
-            seedUITestLoginStateIfNeeded()
-            preloadLastServerURL()
+            viewModel.seedUITestLoginStateIfNeeded()
+            viewModel.preloadLastServerURL()
         }
         .onDisappear {
-            deviceFlowTask?.cancel()
+            viewModel.deviceFlowTask?.cancel()
         }
         .onTapGesture {
-            hideKeyboard()
+            viewModel.hideKeyboard()
         }
         .onChange(of: focusedField) { _, newValue in
             // Auto-validate when URL field loses focus
-            if newValue != .url && !serverURL.isEmpty && !serverValidated && !isConnecting {
-                Task { await validateServer() }
+            if newValue != .url && !viewModel.serverURL.isEmpty && !viewModel.serverValidated && !viewModel.isConnecting {
+                Task { await viewModel.validateServer() }
             }
         }
         .onReceive(NotificationCenter.default.publisher(for: .clientTokenPairingCode)) { notification in
             if let code = notification.userInfo?["code"] as? String {
                 Task {
-                    await performClientTokenPairing(code: code)
+                    await viewModel.performClientTokenPairing(code: code)
                 }
             }
         }
-        .sheet(isPresented: $showQRScanner) {
+        .sheet(isPresented: $viewModel.showQRScanner) {
             QRScannerView { code in
-                showQRScanner = false
+                viewModel.showQRScanner = false
                 Task {
-                    await performClientTokenPairing(code: code)
+                    await viewModel.performClientTokenPairing(code: code)
                 }
             }
         }
@@ -209,15 +140,15 @@ struct SetupView: View {
             urlFieldSection
 
             // Version warning banner
-            if isVersionWarning, let title = versionWarningTitle {
-                versionWarningBanner(title: title, details: versionWarningDetails)
+            if viewModel.isVersionWarning, let title = viewModel.versionWarningTitle {
+                versionWarningBanner(title: title, details: viewModel.versionWarningDetails)
             }
 
             // Sign-in options only appear once the server has been checked and is
             // reachable — no point offering a login before we know it works.
-            if serverValidated {
+            if viewModel.serverValidated {
                 // Primary: browser device-flow sign-in
-                if showDeviceFlowPrimary {
+                if viewModel.showDeviceFlowPrimary {
                     deviceFlowSection
                 }
 
@@ -245,15 +176,15 @@ struct SetupView: View {
             SetupFieldLabel(text: "Server URL")
 
             // Field row
-            let urlState: SetupFieldState = connectionError != nil ? .error : (focusedField == .url ? .focused : .normal)
+            let urlState: SetupFieldState = viewModel.connectionError != nil ? .error : (focusedField == .url ? .focused : .normal)
 
             HStack(spacing: 10) {
                 Image(systemName: "link")
                     .font(.system(size: 15))
-                    .foregroundStyle(serverValidated ? .white.opacity(0.55) : .white.opacity(0.4))
+                    .foregroundStyle(viewModel.serverValidated ? .white.opacity(0.55) : .white.opacity(0.4))
                     .frame(width: 20)
 
-                TextField("", text: $serverURL, prompt: Text(verbatim: "https://romm.example.com")
+                TextField("", text: $viewModel.serverURL, prompt: Text(verbatim: "https://romm.example.com")
                     .foregroundColor(.white.opacity(0.3)))
                     .font(.system(size: 14.5))
                     .foregroundStyle(.white)
@@ -263,23 +194,23 @@ struct SetupView: View {
                     .focused($focusedField, equals: .url)
                     .onSubmit {
                         Task {
-                            await validateServer()
+                            await viewModel.validateServer()
                             focusedField = .username
                         }
                     }
-                    .onChange(of: serverURL) { _, _ in
-                        if serverValidated || isVersionWarning || connectionError != nil {
-                            resetServerValidation()
+                    .onChange(of: viewModel.serverURL) { _, _ in
+                        if viewModel.serverValidated || viewModel.isVersionWarning || viewModel.connectionError != nil {
+                            viewModel.resetServerValidation()
                         }
                     }
 
                 // Version badge (shown when validated or version warning)
-                if (serverValidated || isVersionWarning), let version = detectedServerVersion {
+                if (viewModel.serverValidated || viewModel.isVersionWarning), let version = viewModel.detectedServerVersion {
                     versionBadge(version: version)
                 }
 
                 // Spinner while connecting
-                if isConnecting {
+                if viewModel.isConnecting {
                     ProgressView()
                         .scaleEffect(0.75)
                         .tint(.white.opacity(0.6))
@@ -288,22 +219,22 @@ struct SetupView: View {
             .setupField(urlState)
 
             // Quick input chips — only when not validated and no error/warning
-            if !serverValidated && connectionError == nil && !isVersionWarning {
+            if !viewModel.serverValidated && viewModel.connectionError == nil && !viewModel.isVersionWarning {
                 ScrollView(.horizontal, showsIndicators: false) {
                     HStack(spacing: 6) {
-                        QuickInputChip(text: "192.168.", action: { insertText("192.168.") })
-                        QuickInputChip(text: "http://", action: { insertText("http://") })
-                        QuickInputChip(text: "https://", action: { insertText("https://") })
-                        QuickInputChip(text: ":8080", action: { insertText(":8080") })
-                        QuickInputChip(text: ".com", action: { insertText(".com") })
-                        QuickInputChip(text: ".de", action: { insertText(".de") })
+                        QuickInputChip(text: "192.168.", action: { viewModel.insertText("192.168.") })
+                        QuickInputChip(text: "http://", action: { viewModel.insertText("http://") })
+                        QuickInputChip(text: "https://", action: { viewModel.insertText("https://") })
+                        QuickInputChip(text: ":8080", action: { viewModel.insertText(":8080") })
+                        QuickInputChip(text: ".com", action: { viewModel.insertText(".com") })
+                        QuickInputChip(text: ".de", action: { viewModel.insertText(".de") })
                     }
                     .padding(.vertical, 2)
                 }
             }
 
             // Connection error banner (hard error only — not version warning)
-            if connectionError != nil && !isVersionWarning {
+            if viewModel.connectionError != nil && !viewModel.isVersionWarning {
                 connectionErrorBanner
             }
         }
@@ -313,21 +244,21 @@ struct SetupView: View {
 
     private func versionBadge(version: String) -> some View {
         HStack(spacing: 4) {
-            Image(systemName: isVersionWarning ? "exclamationmark.triangle.fill" : "checkmark")
+            Image(systemName: viewModel.isVersionWarning ? "exclamationmark.triangle.fill" : "checkmark")
                 .font(.system(size: 9, weight: .bold))
             Text("v\(version)")
                 .font(.system(size: 10.5, weight: .semibold))
         }
-        .foregroundStyle(isVersionWarning ? SetupTheme.amber : SetupTheme.greenText)
+        .foregroundStyle(viewModel.isVersionWarning ? SetupTheme.amber : SetupTheme.greenText)
         .padding(.horizontal, 8)
         .padding(.vertical, 4)
         .background(
             Capsule()
-                .fill(isVersionWarning ? SetupTheme.amber.opacity(0.15) : SetupTheme.green.opacity(0.15))
+                .fill(viewModel.isVersionWarning ? SetupTheme.amber.opacity(0.15) : SetupTheme.green.opacity(0.15))
         )
         .overlay(
             Capsule()
-                .strokeBorder(isVersionWarning ? SetupTheme.amber.opacity(0.4) : SetupTheme.green.opacity(0.4), lineWidth: 1)
+                .strokeBorder(viewModel.isVersionWarning ? SetupTheme.amber.opacity(0.4) : SetupTheme.green.opacity(0.4), lineWidth: 1)
         )
     }
 
@@ -337,7 +268,7 @@ struct SetupView: View {
         VStack(alignment: .leading, spacing: 4) {
             Button {
                 withAnimation(.easeInOut(duration: 0.2)) {
-                    isErrorExpanded.toggle()
+                    viewModel.isErrorExpanded.toggle()
                 }
             } label: {
                 HStack(alignment: .top, spacing: 8) {
@@ -345,7 +276,7 @@ struct SetupView: View {
                         .foregroundStyle(SetupTheme.errorIcon)
                         .font(.system(size: 13))
 
-                    Text(connectionError ?? "")
+                    Text(viewModel.connectionError ?? "")
                         .font(.system(size: 12.5))
                         .foregroundStyle(SetupTheme.errorText)
                         .fixedSize(horizontal: false, vertical: true)
@@ -353,8 +284,8 @@ struct SetupView: View {
 
                     Spacer()
 
-                    if connectionErrorDetails != nil {
-                        Image(systemName: isErrorExpanded ? "chevron.up" : "chevron.down")
+                    if viewModel.connectionErrorDetails != nil {
+                        Image(systemName: viewModel.isErrorExpanded ? "chevron.up" : "chevron.down")
                             .foregroundStyle(.white.opacity(0.4))
                             .font(.caption)
                     }
@@ -362,7 +293,7 @@ struct SetupView: View {
             }
             .buttonStyle(.plain)
 
-            if isErrorExpanded, let details = connectionErrorDetails {
+            if viewModel.isErrorExpanded, let details = viewModel.connectionErrorDetails {
                 Text(details)
                     .font(.caption2)
                     .foregroundStyle(.white.opacity(0.5))
@@ -419,15 +350,15 @@ struct SetupView: View {
         VStack(alignment: .leading, spacing: 8) {
             SetupFieldLabel(text: "Authentication Method")
 
-            Picker("Auth Method", selection: $selectedAuthMethod) {
-                ForEach(otherMethods, id: \.self) { method in
+            Picker("Auth Method", selection: $viewModel.selectedAuthMethod) {
+                ForEach(viewModel.otherMethods, id: \.self) { method in
                     Label(method.displayName, systemImage: method.iconName)
                         .tag(method)
                 }
             }
             .pickerStyle(.segmented)
 
-            Text(selectedAuthMethod.description)
+            Text(viewModel.selectedAuthMethod.description)
                 .font(.caption)
                 .foregroundStyle(.white.opacity(0.5))
         }
@@ -447,7 +378,7 @@ struct SetupView: View {
                         .foregroundStyle(.white.opacity(0.4))
                         .frame(width: 20)
 
-                    TextField("", text: $username, prompt: Text("Username")
+                    TextField("", text: $viewModel.username, prompt: Text("Username")
                         .foregroundColor(.white.opacity(0.3)))
                         .font(.system(size: 14.5))
                         .foregroundStyle(.white)
@@ -464,7 +395,7 @@ struct SetupView: View {
             VStack(alignment: .leading, spacing: 8) {
                 SetupFieldLabel(text: "Password")
 
-                let pwState: SetupFieldState = hasLoginError ? .error : (focusedField == .password ? .focused : .normal)
+                let pwState: SetupFieldState = viewModel.hasLoginError ? .error : (focusedField == .password ? .focused : .normal)
 
                 HStack(spacing: 10) {
                     Image(systemName: "lock")
@@ -473,11 +404,11 @@ struct SetupView: View {
                         .frame(width: 20)
 
                     Group {
-                        if showPassword {
-                            TextField("", text: $password, prompt: Text("Password")
+                        if viewModel.showPassword {
+                            TextField("", text: $viewModel.password, prompt: Text("Password")
                                 .foregroundColor(.white.opacity(0.3)))
                         } else {
-                            SecureField("", text: $password, prompt: Text("Password")
+                            SecureField("", text: $viewModel.password, prompt: Text("Password")
                                 .foregroundColor(.white.opacity(0.3)))
                         }
                     }
@@ -486,13 +417,13 @@ struct SetupView: View {
                     .textContentType(.password)
                     .focused($focusedField, equals: .password)
                     .onSubmit {
-                        Task { await handlePrimaryAction() }
+                        Task { await viewModel.handlePrimaryAction() }
                     }
 
                     Button {
-                        showPassword.toggle()
+                        viewModel.showPassword.toggle()
                     } label: {
-                        Image(systemName: showPassword ? "eye.slash" : "eye")
+                        Image(systemName: viewModel.showPassword ? "eye.slash" : "eye")
                             .font(.system(size: 15))
                             .foregroundStyle(.white.opacity(0.4))
                     }
@@ -501,7 +432,7 @@ struct SetupView: View {
                 .setupField(pwState)
 
                 // Login error hint
-                if hasLoginError, let errorMessage = appViewModel.appData.errorMessage {
+                if viewModel.hasLoginError, let errorMessage = appViewModel.appData.errorMessage {
                     HStack(spacing: 5) {
                         Image(systemName: "exclamationmark.circle.fill")
                             .font(.system(size: 11))
@@ -544,7 +475,7 @@ struct SetupView: View {
             )
 
             Button {
-                showQRScanner = true
+                viewModel.showQRScanner = true
             } label: {
                 HStack {
                     Image(systemName: "qrcode.viewfinder")
@@ -580,7 +511,7 @@ struct SetupView: View {
                         .foregroundStyle(.white.opacity(0.4))
                         .frame(width: 20)
 
-                    TextField("", text: $clientTokenInput, prompt: Text("rmm_...")
+                    TextField("", text: $viewModel.clientTokenInput, prompt: Text("rmm_...")
                         .foregroundColor(.white.opacity(0.3)))
                         .font(.system(size: 13.5, design: .monospaced))
                         .foregroundStyle(.white)
@@ -591,7 +522,7 @@ struct SetupView: View {
                 .setupField(focusedField == .token ? .focused : .normal)
             }
 
-            if let error = clientTokenError {
+            if let error = viewModel.clientTokenError {
                 HStack(spacing: 5) {
                     Image(systemName: "exclamationmark.circle.fill")
                         .font(.system(size: 11))
@@ -610,14 +541,14 @@ struct SetupView: View {
     @ViewBuilder
     private var deviceFlowSection: some View {
         VStack(spacing: 12) {
-            if let info = deviceFlowInit, isDeviceFlowRunning {
+            if let info = viewModel.deviceFlowInit, viewModel.isDeviceFlowRunning {
                 deviceFlowWaitingView(info: info)
             } else {
                 Button {
-                    Task { await startDeviceFlow() }
+                    Task { await viewModel.startDeviceFlow() }
                 } label: {
                     HStack(spacing: 8) {
-                        if isDeviceFlowRunning {
+                        if viewModel.isDeviceFlowRunning {
                             ProgressView()
                                 .tint(SetupTheme.onAccent)
                                 .frame(width: 18, height: 18)
@@ -628,15 +559,15 @@ struct SetupView: View {
                         Text("Sign in with Browser")
                             .font(.system(size: 16, weight: .bold))
                     }
-                    .foregroundStyle(deviceFlowButtonEnabled ? SetupTheme.onAccent : .white.opacity(0.4))
+                    .foregroundStyle(viewModel.deviceFlowButtonEnabled ? SetupTheme.onAccent : .white.opacity(0.4))
                     .frame(maxWidth: .infinity)
                     .frame(height: 54)
                     .background(deviceFlowButtonBackground)
                     .clipShape(RoundedRectangle(cornerRadius: 16))
-                    .shadow(color: deviceFlowButtonEnabled ? SetupTheme.coralLight.opacity(0.4) : .clear, radius: 10, x: 0, y: 4)
+                    .shadow(color: viewModel.deviceFlowButtonEnabled ? SetupTheme.coralLight.opacity(0.4) : .clear, radius: 10, x: 0, y: 4)
                 }
                 .buttonStyle(.plain)
-                .disabled(!deviceFlowButtonEnabled)
+                .disabled(!viewModel.deviceFlowButtonEnabled)
 
                 Text("Approve this device in your browser — no password needed.")
                     .font(.caption)
@@ -645,19 +576,15 @@ struct SetupView: View {
                     .frame(maxWidth: .infinity)
             }
 
-            if let deviceFlowError {
+            if let deviceFlowError = viewModel.deviceFlowError {
                 setupErrorHint(deviceFlowError)
             }
         }
     }
 
-    private var deviceFlowButtonEnabled: Bool {
-        !serverURL.isEmpty && !isBusy
-    }
-
     @ViewBuilder
     private var deviceFlowButtonBackground: some View {
-        if deviceFlowButtonEnabled {
+        if viewModel.deviceFlowButtonEnabled {
             SetupTheme.coralGradient
         } else {
             Color.white.opacity(0.06)
@@ -697,7 +624,7 @@ struct SetupView: View {
 
             HStack(spacing: 10) {
                 Button {
-                    openApprovalBrowser(info: info)
+                    viewModel.openApprovalBrowser(info: info)
                 } label: {
                     Label("Reopen Browser", systemImage: "safari")
                         .font(.system(size: 14, weight: .medium))
@@ -710,7 +637,7 @@ struct SetupView: View {
                 .buttonStyle(.plain)
 
                 Button {
-                    cancelDeviceFlow()
+                    viewModel.cancelDeviceFlow()
                 } label: {
                     Text("Cancel")
                         .font(.system(size: 14, weight: .medium))
@@ -732,18 +659,18 @@ struct SetupView: View {
 
     @ViewBuilder
     private var otherOptionsSection: some View {
-        if showDeviceFlowPrimary {
-            if !otherMethods.isEmpty {
+        if viewModel.showDeviceFlowPrimary {
+            if !viewModel.otherMethods.isEmpty {
                 VStack(spacing: 14) {
                     Button {
-                        withAnimation(.easeInOut(duration: 0.2)) { showOtherOptions.toggle() }
+                        withAnimation(.easeInOut(duration: 0.2)) { viewModel.showOtherOptions.toggle() }
                     } label: {
                         HStack {
                             Text("Other sign-in options")
                                 .font(.system(size: 13, weight: .semibold))
                                 .foregroundStyle(.white.opacity(0.7))
                             Spacer()
-                            Image(systemName: showOtherOptions ? "chevron.up" : "chevron.down")
+                            Image(systemName: viewModel.showOtherOptions ? "chevron.up" : "chevron.down")
                                 .font(.caption)
                                 .foregroundStyle(.white.opacity(0.4))
                         }
@@ -751,7 +678,7 @@ struct SetupView: View {
                     }
                     .buttonStyle(.plain)
 
-                    if showOtherOptions {
+                    if viewModel.showOtherOptions {
                         otherMethodsContent
                             .transition(.opacity.combined(with: .move(edge: .top)))
                     }
@@ -766,11 +693,11 @@ struct SetupView: View {
     @ViewBuilder
     private var otherMethodsContent: some View {
         VStack(spacing: 16) {
-            if otherMethods.count > 1 {
+            if viewModel.otherMethods.count > 1 {
                 authMethodPicker
             }
 
-            if selectedAuthMethod == .classic {
+            if viewModel.selectedAuthMethod == .classic {
                 classicAuthFields
             } else {
                 clientTokenAuthSection
@@ -798,15 +725,15 @@ struct SetupView: View {
 
     private var loginButton: some View {
         Button {
-            Task { await handlePrimaryAction() }
+            Task { await viewModel.handlePrimaryAction() }
         } label: {
             ZStack {
-                if isBusy {
+                if viewModel.isBusy {
                     HStack(spacing: 10) {
                         ProgressView()
                             .tint(SetupTheme.onAccent)
                             .frame(width: 18, height: 18)
-                        Text(isExchangingToken ? "Connecting…" : "Logging in…")
+                        Text(viewModel.isExchangingToken ? "Connecting…" : "Logging in…")
                             .font(.system(size: 16, weight: .bold))
                             .foregroundStyle(SetupTheme.onAccent)
                     }
@@ -817,43 +744,43 @@ struct SetupView: View {
                         Text(buttonLabel)
                             .font(.system(size: 16, weight: .bold))
                     }
-                    .foregroundStyle(canSubmit ? SetupTheme.onAccent : .white.opacity(0.4))
+                    .foregroundStyle(viewModel.canSubmit ? SetupTheme.onAccent : .white.opacity(0.4))
                 }
             }
             .frame(maxWidth: .infinity)
             .frame(height: 54)
             .background(buttonBackground)
             .clipShape(RoundedRectangle(cornerRadius: 16))
-            .shadow(color: canSubmit && !isBusy ? buttonShadowColor.opacity(0.4) : .clear, radius: 10, x: 0, y: 4)
+            .shadow(color: viewModel.canSubmit && !viewModel.isBusy ? buttonShadowColor.opacity(0.4) : .clear, radius: 10, x: 0, y: 4)
         }
         .buttonStyle(.plain)
-        .disabled(!canSubmit)
-        .animation(.easeInOut(duration: 0.2), value: canSubmit)
-        .animation(.easeInOut(duration: 0.2), value: isBusy)
-        .animation(.easeInOut(duration: 0.2), value: isVersionWarning)
+        .disabled(!viewModel.canSubmit)
+        .animation(.easeInOut(duration: 0.2), value: viewModel.canSubmit)
+        .animation(.easeInOut(duration: 0.2), value: viewModel.isBusy)
+        .animation(.easeInOut(duration: 0.2), value: viewModel.isVersionWarning)
     }
 
     private var buttonIcon: String {
-        if isVersionWarning { return "arrow.right" }
-        if selectedAuthMethod == .clientToken { return "key.fill" }
+        if viewModel.isVersionWarning { return "arrow.right" }
+        if viewModel.selectedAuthMethod == .clientToken { return "key.fill" }
         return "rectangle.portrait.and.arrow.right"
     }
 
     private var buttonLabel: String {
-        if hasLoginError { return "Try Again" }
-        if isVersionWarning { return "Login Anyway" }
-        if selectedAuthMethod == .clientToken { return "Connect with Token" }
+        if viewModel.hasLoginError { return "Try Again" }
+        if viewModel.isVersionWarning { return "Login Anyway" }
+        if viewModel.selectedAuthMethod == .clientToken { return "Connect with Token" }
         return "Login"
     }
 
     private var buttonShadowColor: Color {
-        isVersionWarning ? SetupTheme.amber : SetupTheme.coralLight
+        viewModel.isVersionWarning ? SetupTheme.amber : SetupTheme.coralLight
     }
 
     @ViewBuilder
     private var buttonBackground: some View {
-        if canSubmit && !isBusy {
-            if isVersionWarning {
+        if viewModel.canSubmit && !viewModel.isBusy {
+            if viewModel.isVersionWarning {
                 SetupTheme.amberGradient
             } else {
                 SetupTheme.coralGradient
@@ -865,416 +792,6 @@ struct SetupView: View {
                         .strokeBorder(Color.white.opacity(0.1), lineWidth: 1)
                 )
         }
-    }
-
-    // MARK: - Primary Action Handler
-
-    private func handlePrimaryAction() async {
-        if !serverValidated {
-            await validateServer()
-        }
-        // Only proceed to login when server is validated, no hard error, and no unaccepted warning
-        guard serverValidated && connectionError == nil else { return }
-        // If version warning: accept incompatible version and remember it so the
-        // foreground version check doesn't immediately warn about the same version.
-        if isVersionWarning {
-            didAcceptIncompatibleVersion = true
-            if let version = detectedServerVersion {
-                appViewModel.acknowledgeServerVersion(version)
-            }
-        }
-        if selectedAuthMethod == .clientToken {
-            await performClientTokenLogin()
-        } else {
-            await performClassicLogin()
-        }
-    }
-
-    // MARK: - Actions
-
-    private func validateServer() async {
-        hideKeyboard()
-        isConnecting = true
-        connectionError = nil
-        connectionErrorDetails = nil
-        isErrorExpanded = false
-        isVersionWarning = false
-        versionWarningTitle = nil
-        versionWarningDetails = nil
-        didAcceptIncompatibleVersion = false
-        detectedAuthCapability = nil
-
-        do {
-            let version = try await appViewModel.fetchServerVersion(from: serverURL)
-            detectedServerVersion = version
-
-            if appViewModel.isVersionCompatible(version) {
-                serverValidated = true
-                connectionError = nil
-                didAcceptIncompatibleVersion = false
-                await detectAuthenticationMethod()
-            } else if version == "development" {
-                versionWarningTitle = "Development Build"
-                versionWarningDetails = "This server is running a development version of RomM. Compatibility is not guaranteed."
-                isVersionWarning = true
-                serverValidated = true
-                await detectAuthenticationMethod()
-            } else {
-                versionWarningTitle = "Incompatible Server Version"
-                versionWarningDetails = "Server v\(version) is outside the supported range (\(appViewModel.minSupportedServerVersion) – \(appViewModel.maxSupportedServerVersion)). Some features may not work correctly."
-                isVersionWarning = true
-                serverValidated = true
-                await detectAuthenticationMethod()
-            }
-        } catch let error as APIClientError {
-            if case .cloudflareProtection = error {
-                connectionError = "Server is protected by Cloudflare"
-                connectionErrorDetails = "This server is behind Cloudflare protection and cannot be accessed directly."
-                isVersionWarning = false
-            } else {
-                let (message, details) = parseConnectionError(error)
-                connectionError = message
-                connectionErrorDetails = details
-                isVersionWarning = false
-            }
-            serverValidated = false
-            didAcceptIncompatibleVersion = false
-        } catch let error as HeartbeatError {
-            switch error {
-            case .decodingError:
-                connectionError = "Invalid Response"
-                connectionErrorDetails = "The server did not return a valid RomM response. Is this a RomM server?"
-            default:
-                let (message, details) = parseConnectionError(error)
-                connectionError = message
-                connectionErrorDetails = details
-            }
-            isVersionWarning = false
-            serverValidated = false
-            didAcceptIncompatibleVersion = false
-        } catch {
-            let (message, details) = parseConnectionError(error)
-            connectionError = message
-            connectionErrorDetails = details
-            isVersionWarning = false
-            serverValidated = false
-            didAcceptIncompatibleVersion = false
-        }
-
-        isConnecting = false
-    }
-
-    // MARK: - Auth Detection
-
-    private func detectAuthenticationMethod() async {
-        let heartbeatRepo = HeartbeatRepository()
-        Logger.auth.info("Detecting authentication methods...")
-        let capabilities = await heartbeatRepo.detectAuthCapabilities(serverURL: serverURL)
-        detectedAuthCapability = capabilities
-        Logger.auth.info("Detected capabilities: \(capabilities.description)")
-        connectionError = nil
-        connectionErrorDetails = nil
-
-        if capabilities.unreachable {
-            serverValidated = false
-            connectionError = "Server is unreachable"
-            connectionErrorDetails = "Could not connect to the server."
-            return
-        }
-        if capabilities.cloudflareBlocked {
-            serverValidated = false
-            connectionError = "Server is protected by Cloudflare"
-            connectionErrorDetails = "This server is behind Cloudflare protection and cannot be accessed directly."
-            return
-        }
-        serverValidated = true
-        if capabilities.classic {
-            selectedAuthMethod = .classic
-        } else if capabilities.clientTokens {
-            selectedAuthMethod = .clientToken
-        }
-    }
-
-    private func parseConnectionError(_ error: Error) -> (message: String, details: String?) {
-        if let apiError = error as? APIClientError {
-            switch apiError {
-            case .cloudflareProtection:
-                return ("Server is protected by Cloudflare", "This server is behind Cloudflare protection and cannot be accessed directly.")
-            case .authenticationRequired:
-                return ("Authentication failed", "Invalid username or password")
-            case .invalidURL(let url):
-                return ("Invalid URL", "The URL '\(url)' is not valid")
-            case .noConfiguration:
-                return ("Configuration error", "Server configuration is missing")
-            case .noCredentials:
-                return ("Credentials missing", "Please provide username and password")
-            case .invalidResponse(let code, _):
-                return ("Server error (\(code))", "The server returned an error response")
-            case .decodingError:
-                return ("Invalid response", "The server did not return a valid RomM response")
-            case .networkError(let underlyingError):
-                return parseGeneralConnectionError(underlyingError)
-            }
-        }
-        return parseGeneralConnectionError(error)
-    }
-
-    private func parseGeneralConnectionError(_ error: Error) -> (message: String, details: String?) {
-        let errorString = error.localizedDescription
-
-        if errorString.contains("certificate") || errorString.contains("SSL") || errorString.contains("TLS") {
-            return ("SSL/TLS certificate error", "Check if the server has a valid certificate or use http:// instead of https://")
-        }
-        if errorString.contains("Could not connect") || errorString.contains("Connection refused") {
-            return ("Connection failed", "Check if the server is running and the URL is correct.")
-        }
-        if errorString.contains("timed out") || errorString.contains("timeout") {
-            return ("Connection timed out", "The server did not respond in time. Check if the server is reachable.")
-        }
-        if errorString.contains("host") || errorString.contains("DNS") || errorString.contains("resolve") {
-            return ("Server not found", "DNS resolution failed. Check the server URL.")
-        }
-        if errorString.contains("network") || errorString.contains("internet") {
-            return ("Network error", "Check your internet connection.")
-        }
-        if errorString.contains("decode") || errorString.contains("JSON") || errorString.contains("invalid") {
-            return ("Invalid response", "The server did not return a valid RomM response. Is this a RomM server?")
-        }
-        return ("Connection error", errorString)
-    }
-
-    private func resetServerValidation() {
-        serverValidated = false
-        detectedServerVersion = nil
-        connectionError = nil
-        connectionErrorDetails = nil
-        isErrorExpanded = false
-        isVersionWarning = false
-        versionWarningTitle = nil
-        versionWarningDetails = nil
-        didAcceptIncompatibleVersion = false
-        // Abort any in-flight browser pairing — it's tied to the old server.
-        cancelDeviceFlow()
-        // NOTE: username/password are intentionally NOT cleared here
-    }
-
-    private func performClassicLogin() async {
-        if let version = detectedServerVersion {
-            appViewModel.saveServerVersion(version)
-        }
-        await appViewModel.saveConfiguration(
-            serverURL: serverURL,
-            username: username,
-            password: password
-        )
-    }
-
-    // MARK: - Client Token Login
-
-    @MainActor
-    private func performClientTokenLogin() async {
-        isExchangingToken = true
-        clientTokenError = nil
-        let service = ClientTokenAuthService()
-        do {
-            let tokenInfo = try await service.validateToken(serverURL: serverURL, token: clientTokenInput)
-            try service.saveToken(clientTokenInput, info: tokenInfo)
-            let setupRepo = SetupRepository()
-            if let version = detectedServerVersion {
-                appViewModel.saveServerVersion(version)
-            }
-            try setupRepo.saveClientTokenSetup(
-                serverURL: serverURL,
-                tokenName: tokenInfo.name,
-                version: detectedServerVersion ?? "unknown",
-                allowIncompatibleVersionLogin: didAcceptIncompatibleVersion
-            )
-            Logger.auth.info("Client token login complete")
-            appViewModel.appData.isLoading = false
-            appViewModel.appData.errorMessage = nil
-            await appViewModel.checkInitialState()
-        } catch {
-            Logger.auth.error("Client token login failed: \(error)")
-            clientTokenError = error.localizedDescription
-        }
-        isExchangingToken = false
-    }
-
-    @MainActor
-    private func performClientTokenPairing(code: String) async {
-        isExchangingToken = true
-        clientTokenError = nil
-        let service = ClientTokenAuthService()
-        do {
-            let (token, tokenInfo) = try await service.exchangeCode(serverURL: serverURL, code: code)
-            try service.saveToken(token, info: tokenInfo)
-            let setupRepo = SetupRepository()
-            if let version = detectedServerVersion {
-                appViewModel.saveServerVersion(version)
-            }
-            try setupRepo.saveClientTokenSetup(
-                serverURL: serverURL,
-                tokenName: tokenInfo.name,
-                version: detectedServerVersion ?? "unknown",
-                allowIncompatibleVersionLogin: didAcceptIncompatibleVersion
-            )
-            Logger.auth.info("Client token pairing complete")
-            appViewModel.appData.isLoading = false
-            appViewModel.appData.errorMessage = nil
-            await appViewModel.checkInitialState()
-        } catch {
-            Logger.auth.error("Client token pairing failed: \(error)")
-            clientTokenError = error.localizedDescription
-        }
-        isExchangingToken = false
-    }
-
-    // MARK: - Device Flow Actions
-
-    @MainActor
-    private func startDeviceFlow() async {
-        hideKeyboard()
-        deviceFlowError = nil
-
-        if !serverValidated {
-            await validateServer()
-        }
-        guard serverValidated && connectionError == nil else { return }
-
-        if isVersionWarning {
-            didAcceptIncompatibleVersion = true
-            if let version = detectedServerVersion {
-                appViewModel.acknowledgeServerVersion(version)
-            }
-        }
-
-        // Server validated but doesn't offer the browser flow — point the user
-        // at the fallback methods instead of failing silently.
-        if let cap = detectedAuthCapability, !cap.deviceFlow {
-            deviceFlowError = "This server doesn't support browser sign-in. Use another method below."
-            withAnimation { showOtherOptions = true }
-            return
-        }
-
-        isDeviceFlowRunning = true
-        let service = DeviceAuthService()
-        do {
-            let info = try await service.initPairing(serverURL: serverURL)
-            deviceFlowInit = info
-            openApprovalBrowser(info: info)
-            deviceFlowTask = Task { await pollDeviceFlow(service: service, info: info) }
-        } catch {
-            isDeviceFlowRunning = false
-            deviceFlowInit = nil
-            deviceFlowError = error.localizedDescription
-        }
-    }
-
-    private func openApprovalBrowser(info: DeviceAuthInitResponse) {
-        guard let url = DeviceAuthService().verificationURL(serverURL: serverURL, init: info) else { return }
-        UIApplication.shared.open(url)
-    }
-
-    @MainActor
-    private func pollDeviceFlow(service: DeviceAuthService, info: DeviceAuthInitResponse) async {
-        do {
-            let token = try await service.pollForToken(serverURL: serverURL, init: info)
-            await finishDeviceFlow(token: token)
-        } catch {
-            if case DeviceAuthError.cancelled = error {
-                // User cancelled — no error message needed.
-            } else {
-                deviceFlowError = error.localizedDescription
-            }
-            isDeviceFlowRunning = false
-            deviceFlowInit = nil
-        }
-    }
-
-    @MainActor
-    private func finishDeviceFlow(token: DeviceAuthTokenResponse) async {
-        let clientService = ClientTokenAuthService()
-        let info = ClientTokenInfo(
-            tokenId: 0,
-            name: "Browser Sign-In",
-            scopes: token.scopes,
-            expiresAt: Self.parseISODate(token.expiresAt)
-        )
-        do {
-            try clientService.saveToken(token.accessToken, info: info)
-            // The device-flow token is bound to a server device — reuse its id
-            // for the save-sync device layer so we don't register twice.
-            UserDefaults.standard.set(token.deviceId, forKey: "sync.deviceId")
-
-            let setupRepo = SetupRepository()
-            if let version = detectedServerVersion {
-                appViewModel.saveServerVersion(version)
-            }
-            try setupRepo.saveClientTokenSetup(
-                serverURL: serverURL,
-                tokenName: info.name,
-                version: detectedServerVersion ?? "unknown",
-                allowIncompatibleVersionLogin: didAcceptIncompatibleVersion
-            )
-            Logger.auth.info("Browser device-flow login complete")
-            appViewModel.appData.isLoading = false
-            appViewModel.appData.errorMessage = nil
-            isDeviceFlowRunning = false
-            deviceFlowInit = nil
-            await appViewModel.checkInitialState()
-        } catch {
-            Logger.auth.error("Failed to store device-flow token: \(error)")
-            deviceFlowError = error.localizedDescription
-            isDeviceFlowRunning = false
-            deviceFlowInit = nil
-        }
-    }
-
-    private func cancelDeviceFlow() {
-        deviceFlowTask?.cancel()
-        deviceFlowTask = nil
-        isDeviceFlowRunning = false
-        deviceFlowInit = nil
-        deviceFlowError = nil
-    }
-
-    private static func parseISODate(_ string: String?) -> Date? {
-        guard let string else { return nil }
-        let formatter = ISO8601DateFormatter()
-        formatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-        if let date = formatter.date(from: string) { return date }
-        formatter.formatOptions = [.withInternetDateTime]
-        return formatter.date(from: string)
-    }
-
-    private func preloadLastServerURL() {
-        guard serverURL.isEmpty, !shouldShowLoginForUITests else { return }
-        if let config = SetupRepository().getSetupConfiguration(), !config.serverURL.isEmpty {
-            serverURL = config.serverURL
-        }
-    }
-
-    private func hideKeyboard() {
-        UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
-    }
-
-    private func insertText(_ text: String) {
-        serverURL += text
-    }
-
-    private func seedUITestLoginStateIfNeeded() {
-        guard shouldShowLoginForUITests else { return }
-
-        if serverURL.isEmpty { serverURL = "https://demo.romm.app" }
-        if username.isEmpty { username = "snapshot-user" }
-        if password.isEmpty { password = "snapshot-password" }
-
-        detectedServerVersion = "3.0.0"
-        serverValidated = true
-        connectionError = nil
-        connectionErrorDetails = nil
-        isVersionWarning = false
-        versionWarningTitle = nil
     }
 }
